WIP: remove 'if', assertions, replace with if-arch
