--- examples/options.yaml
@@ -1,106 +1,0 @@
-package:
-  name: curl
-  version: 7.87.0
-  epoch: 3
-  description: "URL retrieval utility and library"
-  copyright:
-    - license: MIT
-
-environment:
-  contents:
-    repositories:
-      - https://packages.wolfi.dev/os
-    keyring:
-      - https://packages.wolfi.dev/os/wolfi-signing.rsa.pub
-    packages:
-      - brotli-dev
-      - build-base
-      - busybox
-      - ca-certificates-bundle
-      - nghttp2-dev
-      - openssl-dev
-      - wolfi-base
-      - zlib-dev
-
-vars:
-  with-openssl: --with-openssl
-  with-rustls: --without-rustls
-
-options:
-  rustls:
-    vars:
-      with-openssl: --without-openssl
-      with-rustls: --with-rustls
-    environment:
-      contents:
-        packages:
-          add:
-            - rustls-ffi
-          remove:
-            - openssl-dev
-
-pipeline:
-  - uses: fetch
-    with:
-      uri: https://curl.se/download/curl-${{package.version}}.tar.xz
-      expected-sha256: ee5f1a1955b0ed413435ef79db28b834ea5f0fb7c8cfb1ce47175cc3bee08fff
-
-  - if: ${{options.rustls.enabled}} == 'true'
-    runs: |
-      echo "Building with RUSTLS backend"
-
-  - uses: autoconf/configure
-    with:
-      opts: |
-        --enable-ipv6 \
-        --enable-unix-sockets \
-        ${{vars.with-openssl}} \
-        ${{vars.with-rustls}} \
-        --with-nghttp2 \
-        --with-pic \
-        --disable-ldap \
-        --without-libssh2
-
-  - uses: autoconf/make
-
-  - uses: autoconf/make-install
-
-  - uses: strip
-
-subpackages:
-  - name: "curl-dev"
-    description: "headers for libcurl"
-    pipeline:
-      - uses: split/dev
-    dependencies:
-      runtime:
-        - libcurl4
-
-  - name: "curl-doc"
-    description: "documentation for curl"
-    pipeline:
-      - uses: split/manpages
-
-  - if: ${{options.rustls.enabled}} == 'false'
-    name: "libcurl-openssl4"
-    description: "curl library (openssl backend)"
-    dependencies:
-      provides:
-        - libcurl4=7.87.1
-      provider-priority: 5
-    pipeline:
-      - runs: |
-          mkdir -p "${{targets.subpkgdir}}"/usr/lib
-          mv "${{targets.destdir}}"/usr/lib/libcurl.so.* "${{targets.subpkgdir}}"/usr/lib/
-
-  - if: ${{options.rustls.enabled}} == 'true'
-    name: "libcurl-rustls4"
-    description: "curl library (rustls backend)"
-    dependencies:
-      provides:
-        - libcurl4=7.87.1
-      provider-priority: 10
-    pipeline:
-      - runs: |
-          mkdir -p "${{targets.subpkgdir}}"/usr/lib
-          mv "${{targets.destdir}}"/usr/lib/libcurl.so.* "${{targets.subpkgdir}}"/usr/lib/
